requirements.txt
tensorflow-gpu>=1.4.1
urllib3>=1.22
